Refactor background worker by removing rate_unrated_reactions function and reorganizing imports
* Eliminate the rate_unrated_reactions function to streamline the codebase and improve maintainability.
* Update imports to reflect the removal of unused dependencies and introduce a new module for functions.
